# go-server-server-generated/go/model_common_fault.go
/*
 * SystemK
 *
 * API <br/>  Documentation:    <ul>      <li><a href=\"https://atlassian.net/wiki/</a></li>      <li><a href=\"https://git.net/f\">Git</a></li>      <li><a href=\"https://humansinc.atlassian.net/browse/ORM\">Jira</a></li>    </ul>  <br/>  <a href=\"git .yaml\">API Artifact</a><br/>
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type CommonFault struct {

	Code int64 `json:"code"`

	Message string `json:"message,omitempty"`

	Service string `json:"service,omitempty"`

	Operation string `json:"operation,omitempty"`

	Product string `json:"product,omitempty"`

	Cause *interface{} `json:"cause,omitempty"`
}