# go-server-server-generated/go/model_user_params.go
/*
 * SystemK
 *
 * API <br/>  Documentation:    <ul>      <li><a href=\"https://atlassian.net/wiki/</a></li>      <li><a href=\"https://git.net/f\">Git</a></li>      <li><a href=\"https://humansinc.atlassian.net/browse/ORM\">Jira</a></li>    </ul>  <br/>  <a href=\"git .yaml\">API Artifact</a><br/>
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type UserParams struct {
	// логин
	Username string `json:"username,omitempty"`
	// Фамилия*
	FirstName string `json:"firstName,omitempty"`
	// Имя*
	LastName string `json:"lastName,omitempty"`
	// Отчество
	SecondName string `json:"secondName,omitempty"`
	// дата рождения unix time в микросекундах
	BirhDate int32 `json:"birhDate,omitempty"`
	// Электронная почта*
	Email string `json:"email,omitempty"`
	// Пароль храним как хеш
	Password string `json:"password,omitempty"`
	// телефон
	Phone string `json:"phone,omitempty"`
	// User Status
	UserStatus string `json:"userStatus,omitempty"`
	// enum род деятельности клиента/компании
	D string `json:"d,omitempty"`
	// Должность*
	Dd string `json:"dd,omitempty"`
	// Ссылка на инстанс согласия в s3
	S3 string `json:"s3,omitempty"`
	// Ссылка для юр.лица
	S32 string `json:"s32,omitempty"`
}